requirements.txt
APScheduler==3.9.1
matplotlib==3.5.2
Pillow==9.4.0
requests==2.28.0
requests_toolbelt==0.10.1
tqdm==4.64.0
websockets==10.4
psutil==5.9.4
colorama
art